requirements.txt
Flask==3.0.0
pymongo==4.6.0
gunicorn==21.2.0
python-dotenv==1.0.0
pytz==2024.1
Werkzeug==3.0.1
fpdf2==2.7.8
python-docx==1.1.0
redis==5.0.1
pytest==8.2.0
pytest-cov==5.0.0
pytest-flask==1.3.0
pyotp==2.9.0
qrcode[pil]==7.4.2
sentry-sdk[flask]==1.40.6
PyJWT==2.8.0
cryptography==41.0.7
flasgger==0.9.7.1
